Changed id generation strategy for Entity classes
Entity ids are from now on not managed by Hibernate, but generated with
UUID java class and assigned immediately during Entity instantiation.
Drawback: It may sometimes be necessary for Hibernate to do an additinal
select query to see if an Entity does not already exist in the database.
Previously it was able to distinguish based on the fact that an unsaved
Entity had a null id, which is no longer a case.