Cross-site cookies headers handling added
Due to browser changes in 2020 all cross-site cookies are now required
to define SameSite header. If it is not present default SameSite=Lax is
used. Such option causes cross-site cookies to be blocked and so
JSESSIONID cookie and other were blocked making it impossible to log in.
Every Set-Cookie header is now filtered in SecureCookieFilter and
SameSite=None and Secure headers are added to allow them work
cross-site. Secure header is required with SameSite=None, but if the
cookie is sent via unsecured http channel it's getting blocked anyway.
So if the server and client are running on localhost the
SecureCookieFilter should be turned off.